+++ changelog.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Modified the way to obtain audit score
+- Fixed different KeyError
+- Fixed errors in check nodes when name not resolve
+
+## [1.6.1] - 2021-05-09
+### Fixed
+- Fixed error in Satellite statistics produced by API changes
 
 ## [1.6.0] - 2020-12-27
 ### Added